Rename headers for clarity and consistency
